split up page recipes into separate files
