+++ test/source/tests/tests/decrypt.ts
@@ -2,9 +2,10 @@ import { TestUrls } from './../../browser/test_urls';
 import * as ava from 'ava';
 import { Config, Util, TestVariant } from '../../util';
 import { BrowserRecipe } from '../browser_recipe';
-import { SettingsPageRecipe, InboxPageRecipe } from '../page_recipe';
 import { TestWithNewBrowser, TestWithGlobalBrowser } from '../../test';
 import { expect } from "chai";
+import { SettingsPageRecipe } from '../page_recipe/settings-page-recipe';
+import { InboxPageRecipe } from '../page_recipe/inbox-page-recipe';
 
 // tslint:disable:no-blank-lines-func
 // tslint:disable:max-line-length
